Fix many code warnings
* CPakDataChunk: take size_t as size argument and narrow it down to signed int for size types (this gets stored like that in the pak file).
* Starpak offsets now take UINT64_MAX instead of -1 as default argument.
* Additional explicit type casting and shadowing warnings fixed.

+++ src/logic/pakfile.h
@@ -81,8 +81,8 @@ class CPakDataChunk
 
 public:
 	CPakDataChunk() = default;
-	CPakDataChunk(int size, char* data) : size(size), pData(data) {};
-	CPakDataChunk(int pageIndex, int pageOffset, int size, char* data) : pageIndex(pageIndex), pageOffset(pageOffset), size(size), pData(data) {};
+	CPakDataChunk(size_t size, char* data) : size((int)size), pData(data) {};
+	CPakDataChunk(int pageIndex, int pageOffset, size_t size, char* data) : pageIndex(pageIndex), pageOffset(pageOffset), size((int)size), pData(data) {};
 
 private:
 	int pageIndex;
@@ -93,7 +93,7 @@ private:
 
 public:
 
-	inline PagePtr_t GetPointer(int offset=0) { return { pageIndex, pageOffset + offset }; };
+	inline PagePtr_t GetPointer(size_t offset=0) { return { pageIndex, static_cast<int>(pageOffset + offset) }; };
 
 	int GetIndex() { return pageIndex; };
 	char* Data() { return pData; };
@@ -132,7 +132,7 @@ public:
 	inline size_t GetNumPages() const { return m_vPages.size(); };
 
 	inline uint32_t GetVersion() const { return m_Header.fileVersion; }
-	inline void SetVersion(uint32_t version) { m_Header.fileVersion = version; }
+	inline void SetVersion(int version) { m_Header.fileVersion = (short)version; }
 
 	inline void SetStarpakPathsSize(uint16_t len, uint16_t optLen)
 	{
@@ -195,9 +195,9 @@ public:
 	void GenerateFileRelations();
 	void GenerateGuidData();
 
-	CPakPage& FindOrCreatePage(int flags, int alignment, int newDataSize);
+	CPakPage& FindOrCreatePage(int flags, int alignment, size_t newDataSize);
 
-	CPakDataChunk CreateDataChunk(int size, int flags, int alignment);
+	CPakDataChunk CreateDataChunk(size_t size, int flags, int alignment);
 	//_vseginfo_t CreateNewSegment(int size, uint32_t flags, uint32_t alignment, uint32_t vsegAlignment = -1);
 	CPakVSegment& FindOrCreateSegment(int flags, int alignment);
